rename datatypeloader package to datatype
